Rename class and promote field
